requirements.txt
openai>=1.0.0
pydantic>=2.0.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
colorama>=0.4.6
tabulate>=0.9.0
reportlab>=4.0.0
python-dateutil>=2.8.0
